refactor(admin): 상품추가 오류 수정
